Attempt for DFT

+++ Quantisation.py
@@ -31,11 +31,11 @@ Q_50_C = np.array([[17,18,24,47,99,99,99,99],
                     [99,99,99,99,99,99,99,99]])
 
 def Quant_Y(block):
-    return np.round(np.divide(block,Q_50_Y)).astype('int8')
+    return np.round(np.divide(block,Q_50_Y))
 
 
 def Quant_C(block):
-    return np.round(np.divide(block,Q_50_C)).astype('int8')
+    return np.round(np.divide(block,Q_50_C))
 
 def inv_Quant_Y(block):
     return np.multiply(block,Q_50_Y)
